[IMP] orm, base: use bigint for new databases, keep existing ones as is
The "id" primary keys and the foreign keys referencing them are created as
32-bit integers (int4/SERIAL). A long-lived database with high write churn can
exhaust the 2^31-1 space, after which any INSERT on the affected table fails.

This is not theoretical: in May 2025, quay.io (Red Hat's container registry)
suffered a major outage because the primary key of one of its tables reached
the maximum value of a 32-bit integer. See "Quay.io down: primary key has
reached maximum value" https://news.ycombinator.com/item?id=43972181

Odoo already relies on 64-bit sequences without saying so. Since PostgreSQL 10
(October 2017) a sequence carries a type, given by the AS clause; bigint is the
default when the clause is omitted, and SERIAL is what asks for an int4 one. So
in any existing database the sequences that ir.sequence creates for its
standard implementation, with a plain CREATE SEQUENCE, are already bigint,
while the sequences behind the "id" columns are capped at 2^31-1. This change
removes that asymmetry for the databases created from now on.

An earlier attempt (odoo#59809) was refused because it converted existing
databases: rewriting every table and every index of a live database is not
something an update should do. This takes the opposite approach and never
converts anything.

The width is decided per database and read from the database itself:

- databases created from now on use int8 identifiers: "id" columns are
  BIGSERIAL, many2one, many2one_reference and many2many relation columns are
  bigint, and so are the bootstrap tables of base_data.sql, the registry
  signaling tables and res.users.apikeys;

- databases that already exist keep their int4 identifiers. The registry reads
  the width from res_partner.id, a table created by base_data.sql and therefore
  present in every database since its creation, and exposes it as
  Registry.id_column_type. Every column holding an identifier is created with
  that width;

- an identifier column whose width differs from the expected one is left
  untouched instead of being converted, so a database converted by hand,
  entirely or in part, is never rewritten by an update either.

The width cannot be a property of the field: fields are shared between the
registries of a process, so it is resolved through Field.db_column_type(model),
which the schema layer uses to create and to compare columns. Fields that hold
an identifier are marked with Field._id_column; company-dependent and
translated fields keep the jsonb they are stored as.

Some fields store the id of a record without declaring a relation to it, so
they are plain Integer columns: the res_id of a wizard, the id kept after the
record it points to is deleted, or a counter that grows once per record.
Integer therefore takes a `bigint` argument that gives the column the width of
the database, and declares the same column type as a many2one so the value is
not narrowed back to int4 when it is written.

Because identifiers may now exceed the 32-bit range that the XML-RPC
specification defines for <int>, the marshaller sends larger values with the
<i8> extension, as the Apache XML-RPC implementation does. The limit on the
other side belongs to the client: the marshaller of Python's own xmlrpc.client
refuses to send an integer above 2^31, so an external client must emit <i8> as
well, or use JSON-RPC.

The assertions of test_schema, test_search and test_properties follow the width
of the database they run on, and two new tests cover the change: one walks the
installed models and checks that every column holding a record id is as wide as
the ids the database produces and that no foreign key points from one width to
the other, the other covers the detection itself and the fact that an
identifier column is never converted while a regular one still is.

+++ odoo/addons/test_orm/tests/test_properties.py
@@ -1422,7 +1422,7 @@ class PropertiesCase(TestPropertiesMixin):
             UPDATE "test_orm_message"
             SET "attributes" = "__tmp"."attributes"::"jsonb",
                 "write_date" = "__tmp"."write_date"::"timestamp",
-                "write_uid" = "__tmp"."write_uid"::"int4"
+                "write_uid" = "__tmp"."write_uid"::"int8"
             FROM (VALUES %s) AS "__tmp"("id", "attributes", "write_date", "write_uid")
             WHERE "test_orm_message"."id" = "__tmp"."id"
         """]
